new component for navstatus
