+++ src/components/navigation/NavigationStatus.css
@@ -1,0 +1,69 @@
+.nav-status {
+    margin-top: 20px;
+    padding: 15px;
+    border-radius: 8px;
+    width: 80%;
+    max-width: 400px;
+    background-color: #1a1a1a;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+  }
+  
+  .nav-status__header {
+    font-size: 1.2em;
+    font-weight: bold;
+    margin-bottom: 15px;
+    text-align: center;
+    color: #f5f5f5;
+  }
+  
+  .nav-status__content {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    color: #f5f5f5;
+  }
+  
+  .nav-status__coordinates {
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
+    margin-top: 10px;
+    padding-top: 10px;
+    border-top: 1px solid #333;
+  }
+  
+  .nav-status__label--green {
+    color: #4fb185;
+    font-weight: bold;
+  }
+  
+  .nav-status__label--yellow {
+    color: #ffd700;
+    font-weight: bold;
+  }
+  
+  .nav-status__label--red {
+    color: #ff4444;
+    font-weight: bold;
+  }
+  
+  .nav-status__label--gray {
+    color: #888888;
+    font-weight: bold;
+  }
+  
+  .nav-status--green {
+    border: 2px solid #4fb185;
+  }
+  
+  .nav-status--yellow {
+    border: 2px solid #ffd700;
+  }
+  
+  .nav-status--red {
+    border: 2px solid #ff4444;
+  }
+  
+  .nav-status--gray {
+    border: 2px solid #888888;
+  }
